[TASK] ACE-458: Name the delivery tests alike
The five slices that cut the configuration of this branch per component
each brought their own name for the same pair of tests, and the result
was three layouts for one idea:

* Four extensions: "SiteSet/StaticDeliveryTest".
* Two: "Delivery/StaticTemplateDeliveryTest", in a directory of its own.
* Three: "SiteSet/StaticTemplateDeliveryTest", which is what the others
  are now.

Every extension keeps its two halves in one place from here:
"Tests/Functional/SiteSet/StaticTemplateDeliveryTest.php" for the static
template, which runs on both core versions, and
"Tests/Functional/Core13/SiteSet/SiteSetDeliveryTest.php" for the set,
which cannot run on v12 because site sets arrived in TYPO3 v13.1. That
is also the layout of the "main" branch, so a test backported either way
lands where its counterpart already is.

The two fixture folders of the moved tests come with them, and the four
files that named their old path are updated.

Nothing but names and paths changed. Both suites report exactly what
they reported before - 1201 tests on TYPO3 v12 and 1312 on v13 - which
is the point of running them: a rename that loses a test file reports
fewer, and one that duplicates a class reports more.

What is deliberately NOT unified is how the two halves share their
setup: four extensions use an "AbstractDeliveryTestCase", two a
"DeliveryProbeTrait" and three neither. Those are three working designs
rather than three spellings, and picking one is a rewrite of proven
tests, not a rename.
