+++ packages/fgtclb/academic-contact4pages/Tests/Functional/SiteSet/AbstractDeliveryTestCase.php
@@ -15,7 +15,7 @@ use SBUERK\TYPO3\Testing\SiteHandling\SiteBasedTestTrait;
  * The classes are split by core version rather than by mechanism because only one of
  * the two mechanisms exists on both: site sets arrived in TYPO3 v13.1
  * (Feature: #103437), while the static template and the page field `Page TSconfig`
- * work identically on v12 and v13. `StaticDeliveryTest` therefore runs on both
+ * work identically on v12 and v13. `StaticTemplateDeliveryTest` therefore runs on both
  * versions and `Core13\SiteSet\SiteSetDeliveryTest` on v13 only.
  *
  * It matters more here than in the other academic extensions: no seeded development
