+++ packages/fgtclb/academic-jobs/Tests/Functional/SiteSet/DeliveryProbeTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace FGTCLB\AcademicJobs\Tests\Functional\Delivery;
+namespace FGTCLB\AcademicJobs\Tests\Functional\SiteSet;
 
 use FGTCLB\TestingHelper\FunctionalTestCase\FrontendPluginRenderingTrait;
 use SBUERK\TYPO3\Testing\SiteHandling\SiteBasedTestTrait;
@@ -127,7 +127,7 @@ trait DeliveryProbeTrait
                 'clear' => 0,
                 'title' => 'Probe',
                 'constants' => '',
-                'config' => '@import \'EXT:academic_jobs/Tests/Functional/Delivery/Fixtures/Probe.typoscript\'',
+                'config' => '@import \'EXT:academic_jobs/Tests/Functional/SiteSet/Fixtures/Probe.typoscript\'',
                 'include_static_file' => $includeStaticFile,
             ],
         );
